requirements.txt
Flask
SQLAlchemy
Requests
PyYAML
